typ: support equal for recursively nested object types record the info pairs on recursion and treat second encounter as equal
we only have recursion for object fields and only record obj infos.
the second time we see the same comparison we have a recursion loop.
by returning true the loop is broken and we proceed checking other fields.
if it turns out that later fields differ the previous result wont matter.