xelf: add design doc, primarily to organize my own thoughts
i had a bike accident last week due to material fault. the pedal broke and
i landed on face, suffering a light head trauma on the right side making
creative part of the work really hard to do. however the left side of my
brain was fine enough to read a lot and write down this document.