exp: rework reference resolution and change env api
The environments get methods is now limited to the current scope. The new
lookup functions can be used to find a resolver in the current env or any
of its parents.

The next step is to model all expression resolvers as function literals and
use the reference resolution to select the expression resolver.