all: big refactor of parser and drop Tag and Decl, add Named and Raw type
Make position information smaller and add it to Sym and Named for now.
Move type parsing mostly back to the typ package.
The dot naming schema reduces the risk of shadowing. Use common type names
as keywords again, the let and with form will be revisited soon.
With introduction of the Raw to replace the Dyn, we have not much to do in
the parse. Think about parsing as the first resolution stage with context.