exp: change 'as' form to 'con' and the way we declare specs
The new con name is short for construction, conversion or maybe even type
context, but can also work as the word 'con' meaning 'with'.
The old name was hard to use in sentences.

The new way we declare specs uses a xelf type signature string that also
acts a documentation, moves it right to the implementation and allows us
to factor out the common layout code.