std: add new package to move all standard forms into
The resolution context looks up the dyn spec from the environment the first
time it is used and then caches it. If a user needs to set another dyn spec
he can simply clear it before exiting the scope. Restoring the old spec is
not needed because and the resolution context will look it up automatically.