exp, all: lit and typ are no expr elements and must be wrapped as atom now
This change lets us implement the visitor pattern for expression elements,
to reduce type assertion and instead use plain interface calls. The visitor
pattern helps scanning for a particular symbol, for example, without all
the iteration and assertion boiler-plate code for the elements in between.

Another opportunity is that we could add the resolved literal type to atoms
and delay the conversion until we need a typed literal.